requirements.txt
aiogram>=3.0.0
aiohttp
aiosqlite
python-dotenv
requests
beautifulsoup4
lxml
